ED01000 サインイン、OA10000 メニュー作成

+++ backboneAuth-oa1-service/src/main/java/net/jagunma/backbone/auth/authmanager/infra/web/ed01000/Ed01000SignInConverter.java
@@ -55,7 +55,7 @@ public class Ed01000SignInConverter implements SignInRequest {
      *
      * @return 画面表示モード
      */
-    public Integer getMode() {
+    public Short getMode() {
         return vo.getMode();
     }
     /**
